Bump Mojo nightly to dev2026080206
Migrates every OwnedDLHandle.get_function() call site (69 total) to
the new API: parameterized by the return type instead of the full
function-pointer type, and returning a callable now instead of a bare
pointer. This lets FlareRawIO/SimdjsonFFI-style hot-path structs drop
their old ASAP-destruction borrow tricks in favor of resolving the
symbol fresh per call against a long-lived handle field.

Also fixes, across the whole tree:
  - `String(unsafe_from_utf8=x.as_bytes()[...])` assigned back onto
    the same binding it reads from (now flagged as aliasing) — routed
    through a temporary in each case.
  - The pointer-arithmetic sugar removal (`ptr[i]` → `ptr[unsafe_offset=i]`,
    `ptr + i` → `ptr.unsafe_offset(i)`) and the accompanying method
    renames (bitcast, load, free, memcpy, init_pointee_copy/move,
    destroy_pointee, take_pointee, the `read` parameter convention,
    `__del__` → `__deinit__`) across ~1600 call sites.
  - `Span` relocated from `std.memory` to `std.collections`.
  - `os.path.remove` → `os.remove`.
  - Assorted `Int`/`UInt` comparison fallout from `perf_counter_ns()`
    and `monotonic()` now returning `Int`.

+++ flare/http/reliability.mojo
@@ -269,6 +269,6 @@ struct PostHocDeadline[Inner: Handler & Copyable & Defaultable](
         var start = perf_counter_ns()
         var resp = self.inner.serve(req)
         var elapsed_ms = (perf_counter_ns() - start) // 1_000_000
-        if elapsed_ms > UInt(self.budget_ms):
+        if elapsed_ms > self.budget_ms:
             return Response(status=504, reason=String("Gateway Timeout"))
         return resp^
